+++ .github/workflows/pipeline.yml
@@ -191,37 +191,6 @@ jobs:
               | xargs -0 --no-run-if-empty pnpm prettier --check --
           fi
 
-  tests-eslint-plugin:
-    runs-on: blacksmith-4vcpu-ubuntu-2404
-    needs:
-      - pre-job
-    if: needs.pre-job.outputs.should_skip != 'true'
-    steps:
-      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
-        with:
-          persist-credentials: false
-      - uses: pnpm/action-setup@0e279bb959325dab635dd2c09392533439d90093 # v6.0.8
-        with:
-          version: 11.4.0
-      - name: Use Node.js ${{ env.NODE_VERSION }} without cache
-        if: env.CI_CACHE_ALLOWED != 'true'
-        uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
-        with:
-          node-version: ${{ env.NODE_VERSION }}
-          package-manager-cache: false
-      - name: Use Node.js ${{ env.NODE_VERSION }} with pnpm cache
-        if: env.CI_CACHE_ALLOWED == 'true'
-        uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # zizmor: ignore[cache-poisoning] eslint-plugin test dependency cache only; no published artifacts are built from this cached state
-        with:
-          node-version: ${{ env.NODE_VERSION }}
-          cache: "pnpm"
-          cache-dependency-path: "pnpm-lock.yaml"
-      - name: install dependencies
-        run: |
-          pnpm install
-      - name: run eslint-plugin tests
-        run: pnpm --filter @repo/eslint-plugin run test
-
   tests-shared:
     timeout-minutes: 15
     runs-on: blacksmith-4vcpu-ubuntu-2404
@@ -808,7 +777,6 @@ jobs:
       [
         lint,
         prettier-check,
-        tests-eslint-plugin,
         tests-shared,
         tests-web,
         tests-worker,
