+++ .vscode/settings.json
@@ -1,21 +1,12 @@
 {
   "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": "explicit"
+    "source.fixAll.oxc": "explicit"
   },
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   "editor.formatOnSave": true,
   "editor.rulers": [100],
   "editor.tabSize": 2,
 
-  "eslint.validate": [
-    "javascript",
-    "javascriptreact",
-    "astro",
-    "typescript",
-    "typescriptreact"
-  ],
-
-  "eslint.rules.customizations": [{ "rule": "*", "severity": "warn" }],
   "typescript.tsdk": "node_modules/typescript/lib",
   "prettier.documentSelectors": [
     "**/*.{cjs,mjs,ts,tsx,astro,md,mdx,json,yaml,yml}"
@@ -26,13 +17,5 @@
   "docwriter.style": "JSDoc",
   "[prisma]": {
     "editor.defaultFormatter": "Prisma.prisma"
-  },
-  "eslint.lintTask.enable": true,
-  "eslint.workingDirectories": [
-    { "pattern": "./web" },
-    { "pattern": "./worker" },
-    { "pattern": "./packages/shared" },
-    { "pattern": "./ee" }
-  ],
-  "eslint.useFlatConfig": true
+  }
 }
