+++ ee/package.json
@@ -22,8 +22,8 @@
     "build:check": "tsc",
     "typecheck": "dotenv -e ../../.env -- tsgo --noEmit --skipLibCheck --incremental --tsBuildInfoFile .tsbuildinfo",
     "dev": "tsc --watch",
-    "lint": "oxlint . --deny-warnings && node ../scripts/lint/check-oxlint-replacements.mjs --vitest .",
-    "lint:fix": "oxlint . --fix && node ../scripts/lint/check-oxlint-replacements.mjs --vitest ."
+    "lint": "oxlint . --deny-warnings",
+    "lint:fix": "oxlint . --fix"
   },
   "dependencies": {
     "@langfuse/shared": "workspace:*",
@@ -34,11 +34,9 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@types/node": "^24.3.0",
     "@typescript/native-preview": "7.0.0-dev.20260122.3",
-    "eslint": "^9.39.2",
     "eslint-plugin-turbo": "2.9.14",
     "oxlint": "1.67.0",
     "oxlint-tsgolint": "^0.23.0",
