--- packages/config-eslint/index.d.ts
@@ -1,5 +1,0 @@
-import type { Linter } from "eslint";
-
-declare const config: Linter.Config[];
-
-export default config;
