--- packages/config-eslint/index.js
@@ -1,2 +1,0 @@
-// @ts-check
-export { default } from "./base.js";
