--- packages/config-eslint/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@repo/eslint-config",
-  "version": "0.0.0",
-  "license": "MIT",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./index.js",
-    "./base": "./base.js",
-    "./next": "./next.js"
-  },
-  "files": [
-    "index.js",
-    "base.js",
-    "next.js"
-  ],
-  "dependencies": {
-    "@eslint/js": "^9.39.2",
-    "@repo/eslint-plugin": "workspace:*",
-    "eslint-config-next": "16.2.6",
-    "eslint-config-prettier": "^10.1.8",
-    "eslint-config-turbo": "2.9.14",
-    "eslint-plugin-only-warn": "^1.1.0",
-    "eslint-plugin-prettier": "^5.5.4",
-    "globals": "^16.0.0",
-    "typescript-eslint": "^8.57.1"
-  },
-  "peerDependencies": {
-    "eslint": "^9.39.0",
-    "typescript": ">=5.0.0"
-  }
-}
