--- packages/shared/eslint.config.mjs
@@ -1,3 +1,0 @@
-import baseConfig from "@repo/eslint-config";
-
-export default [...baseConfig];
