+++ packages/shared/package.json
@@ -61,8 +61,8 @@
     "build:check": "tsc",
     "typecheck": "dotenv -e ../../.env -- tsc --noEmit --incremental --tsBuildInfoFile .tsbuildinfo",
     "dev": "tsc --watch",
-    "lint": "oxlint . --deny-warnings && node ../../scripts/lint/check-oxlint-replacements.mjs --vitest .",
-    "lint:fix": "oxlint . --fix && node ../../scripts/lint/check-oxlint-replacements.mjs --vitest .",
+    "lint": "oxlint . --deny-warnings",
+    "lint:fix": "oxlint . --fix",
     "test": "dotenv -e ../../.env -- vitest run",
     "db:migrate": "DISABLE_ERD=false dotenv -e ../../.env -- npx prisma migrate dev",
     "db:push": "DISABLE_ERD=false dotenv -e ../../.env -- npx prisma db push",
@@ -136,14 +136,12 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@types/lodash": "^4.17.24",
     "@types/node": "^24.3.0",
     "@types/nodemailer": "^7.0.11",
     "@types/pg": "^8.11.10",
     "@types/react": "19.2.14",
-    "eslint": "^9.39.2",
     "eslint-plugin-turbo": "2.9.14",
     "oxlint": "1.67.0",
     "oxlint-tsgolint": "^0.23.0",
