--- web/eslint.config.mjs
@@ -1,32 +1,0 @@
-import storybook from "eslint-plugin-storybook";
-
-import nextConfig from "@repo/eslint-config/next";
-
-export default [
-  ...nextConfig,
-  ...storybook.configs["flat/recommended"],
-
-  // Restrict react-icons imports
-  {
-    name: "langfuse/web/react-icons-restriction",
-    rules: {
-      "no-restricted-imports": [
-        "error",
-        {
-          patterns: [
-            {
-              regex: "^react-icons$",
-              message:
-                "Only react-icons/si and react-icons/tb are allowed. Please use lucide-react for other icons.",
-            },
-            {
-              regex: "^react-icons/(?!si(?:/|$)|tb(?:/|$)).*",
-              message:
-                "Only react-icons/si and react-icons/tb are allowed. Please use lucide-react for other icons.",
-            },
-          ],
-        },
-      ],
-    },
-  },
-];
