+++ web/package.json
@@ -12,8 +12,8 @@
     "typecheck": "dotenv -e ../.env -- tsgo -p tsconfig.build.json --noEmit --skipLibCheck --incremental --tsBuildInfoFile .tsbuildinfo",
     "dev": "dotenv -e ../.env -- next dev",
     "dev:http": "dotenv -e ../.env -- next dev --experimental-https --experimental-https-key ./localhost+1-key.pem --experimental-https-cert ./localhost+1.pem",
-    "lint": "dotenv -e ../.env -- oxlint . --deny-warnings && node ../scripts/lint/check-oxlint-replacements.mjs --vitest . --tailwind .",
-    "lint:fix": "dotenv -e ../.env -- oxlint . --fix && node ../scripts/lint/check-oxlint-replacements.mjs --vitest . --tailwind .",
+    "lint": "dotenv -e ../.env -- oxlint . --deny-warnings",
+    "lint:fix": "dotenv -e ../.env -- oxlint . --fix",
     "analyze": "dotenv -e ../.env -- next experimental-analyze",
     "clean": "rm -rf node_modules",
     "start": "dotenv -e ../.env -- sh -c 'NEXT_MANUAL_SIG_HANDLE=true next start'",
@@ -173,7 +173,6 @@
   "devDependencies": {
     "@jedmao/location": "^3.0.0",
     "@playwright/test": "^1.57.0",
-    "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@storybook/addon-a11y": "^10.3.6",
     "@storybook/addon-docs": "^10.3.6",
@@ -186,7 +185,6 @@
     "@types/bcryptjs": "^2.4.6",
     "@types/chroma-js": "^3.1.2",
     "@types/cors": "^2.8.19",
-    "@types/eslint": "^9.6.1",
     "@types/lodash": "^4.17.24",
     "@types/node": "24.3.0",
     "@types/react": "19.2.14",
@@ -197,8 +195,6 @@
     "dotenv": "^16.6.1",
     "dotenv-cli": "^7.4.2",
     "dotenv-expand": "^10.0.0",
-    "eslint": "^9.39.2",
-    "eslint-config-next": "16.2.6",
     "eslint-plugin-storybook": "^10.3.6",
     "jsdom": "^26.1.0",
     "node-mocks-http": "^1.17.2",
