--- worker/eslint.config.mjs
@@ -1,11 +1,0 @@
-import baseConfig from "@repo/eslint-config";
-
-export default [
-  ...baseConfig,
-
-  // Worker-specific ignores
-  {
-    name: "langfuse/worker/ignores",
-    ignores: ["**/*test*.*", "**/worker-thread.js"],
-  },
-];
